+++ src/core/converter/other/XplaneFmsToAeroflyFlightConverter.ts
@@ -35,6 +35,11 @@ export class XplaneFmsToAeroflyFlightConverter extends StringToAeroflyFlightConv
             throw new Error("File format only contains one flight plan");
         }
 
+        const version = this.getVersion(content);
+        if (version < 1100) {
+            throw new Error("Flight plan file format too old, please use X-Plane 11/12 FMS");
+        }
+
         const waypoints = this.getWaypoints(content);
 
         const departureRunway = this.getRunway(content, "DEPRWY");
@@ -54,6 +59,15 @@ export class XplaneFmsToAeroflyFlightConverter extends StringToAeroflyFlightConv
         flightplan._missionBriefing = "";
     }
 
+    private getVersion(content: string): number {
+        const match = content.match(/\s(\d+) VERSION\s/i);
+        if (!match || !match[1]) {
+            throw new Error("Missing version information from import file");
+        }
+
+        return this.parseNumberOrError(match[1]);
+    }
+
     private getRunway(content: string, type: "DEPRWY" | "DESRWY"): string | null {
         const match = content.match(new RegExp(`\\s${type} RW(\\S+)`));
         return match ? match[1] : null;
@@ -65,12 +79,14 @@ export class XplaneFmsToAeroflyFlightConverter extends StringToAeroflyFlightConv
             throw new Error("No nav lines found");
         }
         return Array.from(waypointLines).map((m): XplaneFmsWaypoint => {
-            const mString = m.join(" ");
+            if (m.length !== 6) {
+                throw new Error(`Broken waypoint, expected 6, got ${m.length.toString()} cells`);
+            }
             return {
                 identifier: m[2],
-                type: this.parseNumberOrError(m[1], mString) as XplaneFmsWaypointType,
-                lat: this.parseNumberOrError(m[4], mString),
-                lon: this.parseNumberOrError(m[5], mString),
+                type: this.parseNumberOrError(m[1], m[0].trim()) as XplaneFmsWaypointType,
+                lat: this.parseNumberOrError(m[4], m[0].trim()),
+                lon: this.parseNumberOrError(m[5], m[0].trim()),
                 elevationFeet: this.parseNumber(m[3], 0),
             };
         });