+++ src/core/services/ImportFileFinderService.test.ts
@@ -9,7 +9,7 @@ describe("ImportFileFinderService", () => {
         const finder = new ImportFileFinderService(config);
         const files = finder.findImportFiles();
 
-        assert.strictEqual(8, files?.length);
+        assert.strictEqual(11, files?.length);
         //console.log(files);
     });
 });