+++ types/test/loadFixture.d.ts
@@ -1,0 +1,2 @@
+export declare function loadFixture(filename: string): string;
+//# sourceMappingURL=loadFixture.d.ts.map